backend/ai-shopify-agent/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1

# LangChain & LangGraph - Use compatible versions
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
langchain-core>=0.3.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Pydantic
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP & WebSockets
httpx>=0.27.0
websockets>=13.0  # Required for Deepgram streaming STT
requests>=2.31.0
aiohttp>=3.9.0

# Voice Services
elevenlabs>=2.0.0
deepgram-sdk>=3.0.0  # Official Deepgram SDK

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Utilities
python-multipart>=0.0.9
